Separated login to display data from Collectors, added tests
